Loking down mutations (post and comments)

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -99,21 +99,36 @@ const Mutation = {
   },
   async updatePost(parent, args, { prisma, request }, info) {
     const userId = getUserId(request);
-
+    
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+    
+    if (!postExists) {
+      throw new Error("Unable to update this post");
+    }
+    
+    console.log("TCL: updatePost -> userId", userId)
     return prisma.mutation.updatePost({
       where: {
-        id: userId
+        id: args.id
       },
       data: args.data
     }, info);
   },
-  createComment(parent, args, { prisma }, info) {
+  async createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
+    
     return prisma.mutation.createComment({
       data: {
         text: args.data.text,
         author: {
           connect: {
-            id: args.data.author
+            id: userId
           }
         },
         post: {
@@ -124,7 +139,19 @@ const Mutation = {
       }
     }, info)
   },
-  deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!commentExists) {
+      throw new Error("Unable to delete this comment!");
+    }
+
    return prisma.mutation.deleteComment({
     where: {
       id: args.id
@@ -132,7 +159,19 @@ const Mutation = {
    }, info)
 
   },
-  updateComment(parent, args, { prisma }, info){
+  async updateComment(parent, args, { prisma, request }, info){
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!commentExists) {
+      throw new Error("Unable to update this comment!");
+    }
+
     return prisma.mutation.updateComment({
       where: {
         id: args.id
